feature/frontend-cleanup
massive refactor of frontend code, identify and outline architecture and current state
