+++ web/src/app/App.tsx
@@ -6,9 +6,9 @@ import { useAppDispatch } from './hooks'
 import { websocketConnect } from './middleware/websocket'
 import { AuthModal } from '../features/auth/components/AuthModal'
 import { Layout } from '../components/Layout'
-import { ProjectManager } from '../components/ProjectManager'
-import { SessionManager } from '../components/SessionManager'
-import { SessionView } from '../components/SessionView'
+import { ProjectManager } from '../features/projects/components/ProjectManager'
+import { SessionManager } from '../features/sessions/components/SessionManager'
+import { SessionView } from '../features/sessions/components/SessionView'
 import { selectCredentials } from '../features/auth/slice/authSlice'
 import { StoreSync } from '../shared/components/StoreSync'
 import { useAppStore } from '../store'
