+++ web/src/features/projects/components/AddSSHProjectForm.tsx
@@ -1,10 +1,7 @@
 import { useState } from 'react'
-import { useCreateProjectMutation } from '../api/projectsApi'
-import { CreateProjectRequest } from '../../../shared/types/schemas'
-import { z } from 'zod'
-import { ProjectConfigSchema } from '../../../shared/types/schemas'
-
-type ProjectConfig = z.infer<typeof ProjectConfigSchema>
+import { useMutation, useQueryClient } from '@tanstack/react-query'
+import { projectsApi } from '../../../api/client'
+import { CreateProjectRequest, ProjectConfig } from '../../../types'
 
 interface AddSSHProjectFormProps {
   onSuccess?: () => void
@@ -12,7 +9,7 @@ interface AddSSHProjectFormProps {
 }
 
 export function AddSSHProjectForm({ onSuccess, onCancel }: AddSSHProjectFormProps) {
-  const [createProject, { isLoading, error }] = useCreateProjectMutation()
+  const queryClient = useQueryClient()
   const [formData, setFormData] = useState({
     name: '',
     ssh_host: '',
@@ -25,161 +22,153 @@ export function AddSSHProjectForm({ onSuccess, onCancel }: AddSSHProjectFormProp
   })
   const [envVarInput, setEnvVarInput] = useState({ key: '', value: '' })
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault()
+  const createProject = useMutation({
+    mutationFn: async () => {
+      const config: ProjectConfig = {
+        ssh_host: formData.ssh_host,
+        ssh_port: formData.ssh_port,
+        ssh_key_path: formData.ssh_key_path,
+        remote_project_path: formData.remote_project_path,
+        remote_setup_cmd: formData.remote_setup_cmd,
+        environment_vars: formData.environment_vars,
+      }
 
-    const config: Partial<ProjectConfig> = {
-      ssh_host: formData.ssh_host,
-      ssh_port: formData.ssh_port,
-      ssh_key_path: formData.ssh_key_path,
-      remote_project_path: formData.remote_project_path,
-      remote_setup_cmd: formData.remote_setup_cmd,
-      environment_vars: formData.environment_vars,
-    }
+      const request: CreateProjectRequest = {
+        name: formData.name,
+        path: formData.remote_project_path, // Use remote path as the path
+        default_branch: formData.default_branch,
+        setup_command: formData.remote_setup_cmd, // Also store in top-level for compatibility
+      }
 
-    const request: CreateProjectRequest = {
-      name: formData.name,
-      path: formData.remote_project_path,
-      default_branch: formData.default_branch,
-      setup_command: formData.remote_setup_cmd,
-      config: config as ProjectConfig,
-    }
-
-    try {
-      await createProject(request).unwrap()
+      const response = await projectsApi.create({
+        ...request,
+        config,
+      })
+      return response.data
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['projects'] })
       onSuccess?.()
-    } catch (err) {
-      console.error('Failed to create SSH project:', err)
-    }
+    },
+  })
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault()
+    createProject.mutate()
   }
 
   const addEnvironmentVar = () => {
     if (envVarInput.key && envVarInput.value) {
-      setFormData({
-        ...formData,
+      setFormData(prev => ({
+        ...prev,
         environment_vars: {
-          ...formData.environment_vars,
+          ...prev.environment_vars,
           [envVarInput.key]: envVarInput.value,
         },
-      })
+      }))
       setEnvVarInput({ key: '', value: '' })
     }
   }
 
   const removeEnvironmentVar = (key: string) => {
-    const newVars = { ...formData.environment_vars }
-    delete newVars[key]
-    setFormData({ ...formData, environment_vars: newVars })
+    setFormData(prev => {
+      const { [key]: _, ...rest } = prev.environment_vars
+      return {
+        ...prev,
+        environment_vars: rest,
+      }
+    })
   }
+
+  const setupCommandHint = `Example: source ~/.bashrc && cd $WORKTREE_PATH && npm install
+Available variables:
+- $PROJECT_PATH: Local project path
+- $WORKTREE_PATH: Remote worktree path
+- $REMOTE_PROJECT_PATH: Remote project base path
+- Plus any custom environment variables you define`
 
   return (
     <form onSubmit={handleSubmit} className="space-y-4">
       <div>
-        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-          Project Name
-        </label>
+        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">Project Name</label>
         <input
           type="text"
           value={formData.name}
           onChange={(e) => setFormData({ ...formData, name: e.target.value })}
-          className="w-full p-2 border border-gray-300 dark:border-gray-600 rounded-md focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+          className="mt-1 block w-full rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
           required
         />
+      </div>
+
+      <div>
+        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">SSH Host</label>
+        <input
+          type="text"
+          value={formData.ssh_host}
+          onChange={(e) => setFormData({ ...formData, ssh_host: e.target.value })}
+          placeholder="user@hostname"
+          className="mt-1 block w-full rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+          required
+        />
+        <p className="mt-1 text-sm text-gray-500 dark:text-gray-400">Format: user@hostname</p>
       </div>
 
       <div className="grid grid-cols-2 gap-4">
         <div>
-          <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-            SSH Host
-          </label>
-          <input
-            type="text"
-            value={formData.ssh_host}
-            onChange={(e) => setFormData({ ...formData, ssh_host: e.target.value })}
-            className="w-full p-2 border border-gray-300 dark:border-gray-600 rounded-md focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
-            placeholder="example.com"
-            required
-          />
-        </div>
-        <div>
-          <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-            SSH Port
-          </label>
+          <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">SSH Port</label>
           <input
             type="number"
             value={formData.ssh_port}
-            onChange={(e) => setFormData({ ...formData, ssh_port: parseInt(e.target.value) })}
-            className="w-full p-2 border border-gray-300 dark:border-gray-600 rounded-md focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+            onChange={(e) => setFormData({ ...formData, ssh_port: parseInt(e.target.value) || 22 })}
+            className="mt-1 block w-full rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+          />
+        </div>
+
+        <div>
+          <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">SSH Key Path</label>
+          <input
+            type="text"
+            value={formData.ssh_key_path}
+            onChange={(e) => setFormData({ ...formData, ssh_key_path: e.target.value })}
+            className="mt-1 block w-full rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
             required
           />
         </div>
       </div>
 
       <div>
-        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-          SSH Key Path
-        </label>
-        <input
-          type="text"
-          value={formData.ssh_key_path}
-          onChange={(e) => setFormData({ ...formData, ssh_key_path: e.target.value })}
-          className="w-full p-2 border border-gray-300 dark:border-gray-600 rounded-md focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
-          placeholder="~/.ssh/id_rsa"
-          required
-        />
-      </div>
-
-      <div>
-        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-          Remote Project Path
-        </label>
+        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">Remote Project Path</label>
         <input
           type="text"
           value={formData.remote_project_path}
           onChange={(e) => setFormData({ ...formData, remote_project_path: e.target.value })}
-          className="w-full p-2 border border-gray-300 dark:border-gray-600 rounded-md focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
-          placeholder="/home/user/project"
+          placeholder="/home/user/projects/myproject"
+          className="mt-1 block w-full rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
           required
         />
+        <p className="mt-1 text-sm text-gray-500 dark:text-gray-400">Full path to the project on the remote server</p>
       </div>
 
       <div>
-        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-          Default Branch
-        </label>
+        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">Default Branch</label>
         <input
           type="text"
           value={formData.default_branch}
           onChange={(e) => setFormData({ ...formData, default_branch: e.target.value })}
-          className="w-full p-2 border border-gray-300 dark:border-gray-600 rounded-md focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+          className="mt-1 block w-full rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
         />
       </div>
 
       <div>
-        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-          Remote Setup Script
-        </label>
-        <textarea
-          value={formData.remote_setup_cmd}
-          onChange={(e) => setFormData({ ...formData, remote_setup_cmd: e.target.value })}
-          className="w-full p-2 border border-gray-300 dark:border-gray-600 rounded-md focus:ring-2 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100 font-mono text-sm"
-          rows={4}
-          placeholder="#!/bin/bash&#10;# Commands to run on remote server when creating session"
-        />
-      </div>
-
-      <div>
-        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
-          Environment Variables
-        </label>
-        <div className="space-y-2">
+        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">Environment Variables</label>
+        <div className="mt-1 space-y-2">
           {Object.entries(formData.environment_vars).map(([key, value]) => (
             <div key={key} className="flex items-center gap-2">
-              <code className="bg-gray-100 dark:bg-gray-700 text-gray-900 dark:text-gray-100 px-2 py-1 rounded">{key}={value}</code>
+              <span className="text-sm font-mono bg-gray-100 dark:bg-gray-700 text-gray-900 dark:text-gray-100 px-2 py-1 rounded">{key}={value}</span>
               <button
                 type="button"
                 onClick={() => removeEnvironmentVar(key)}
-                className="text-red-600 dark:text-red-400 hover:text-red-700 dark:hover:text-red-300"
+                className="text-red-600 dark:text-red-400 hover:text-red-700 dark:hover:text-red-300 text-sm"
               >
                 Remove
               </button>
@@ -190,20 +179,20 @@ export function AddSSHProjectForm({ onSuccess, onCancel }: AddSSHProjectFormProp
               type="text"
               value={envVarInput.key}
               onChange={(e) => setEnvVarInput({ ...envVarInput, key: e.target.value })}
-              placeholder="Key"
-              className="flex-1 p-2 border border-gray-300 dark:border-gray-600 rounded-md bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+              placeholder="KEY"
+              className="flex-1 rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
             />
             <input
               type="text"
               value={envVarInput.value}
               onChange={(e) => setEnvVarInput({ ...envVarInput, value: e.target.value })}
-              placeholder="Value"
-              className="flex-1 p-2 border border-gray-300 dark:border-gray-600 rounded-md bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+              placeholder="value"
+              className="flex-1 rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
             />
             <button
               type="button"
               onClick={addEnvironmentVar}
-              className="px-4 py-2 bg-gray-100 dark:bg-gray-700 text-gray-700 dark:text-gray-200 rounded-md hover:bg-gray-200 dark:hover:bg-gray-600"
+              className="px-3 py-2 bg-gray-200 dark:bg-gray-700 text-gray-700 dark:text-gray-200 rounded-md hover:bg-gray-300 dark:hover:bg-gray-600"
             >
               Add
             </button>
@@ -211,26 +200,32 @@ export function AddSSHProjectForm({ onSuccess, onCancel }: AddSSHProjectFormProp
         </div>
       </div>
 
-      {error && (
-        <div className="bg-red-50 dark:bg-red-900 border border-red-200 dark:border-red-700 rounded-md p-3">
-          <p className="text-sm text-red-800 dark:text-red-300">Failed to create project</p>
-        </div>
-      )}
+      <div>
+        <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">Remote Setup Command</label>
+        <textarea
+          value={formData.remote_setup_cmd}
+          onChange={(e) => setFormData({ ...formData, remote_setup_cmd: e.target.value })}
+          rows={3}
+          className="mt-1 block w-full rounded-md border-gray-300 dark:border-gray-600 shadow-sm focus:border-blue-500 dark:focus:border-blue-400 focus:ring-blue-500 dark:focus:ring-blue-400 bg-white dark:bg-gray-800 text-gray-900 dark:text-gray-100"
+          placeholder="Commands to run when setting up a new session"
+        />
+        <p className="mt-1 text-sm text-gray-500 dark:text-gray-400 whitespace-pre-wrap">{setupCommandHint}</p>
+      </div>
 
-      <div className="flex justify-end gap-3">
+      <div className="flex justify-end gap-2">
         <button
           type="button"
           onClick={onCancel}
-          className="px-4 py-2 text-gray-700 dark:text-gray-200 bg-gray-100 dark:bg-gray-700 rounded-md hover:bg-gray-200 dark:hover:bg-gray-600"
+          className="px-4 py-2 text-gray-700 dark:text-gray-200 bg-white dark:bg-gray-700 border border-gray-300 dark:border-gray-600 rounded-md hover:bg-gray-50 dark:hover:bg-gray-600"
         >
           Cancel
         </button>
         <button
           type="submit"
-          disabled={isLoading}
-          className="px-4 py-2 bg-blue-600 dark:bg-blue-800 text-white rounded-md hover:bg-blue-700 disabled:opacity-50"
+          disabled={createProject.isPending || !formData.name || !formData.ssh_host || !formData.remote_project_path}
+          className="px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700 disabled:opacity-50"
         >
-          {isLoading ? 'Creating...' : 'Create SSH Project'}
+          {createProject.isPending ? 'Adding...' : 'Add SSH Project'}
         </button>
       </div>
     </form>
