+++ web/src/features/projects/components/ProjectList.tsx
@@ -1,10 +1,10 @@
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { useState } from 'react'
-import { projectsApi } from '../api/client'
-import { useAppStore } from '../store'
-import { Project, CreateProjectRequest } from '../types'
+import { projectsApi } from '../../../api/client'
+import { useAppStore } from '../../../store'
+import { Project, CreateProjectRequest } from '../../../types'
 import { AddSSHProjectForm } from './AddSSHProjectForm'
-import { DropdownMenu } from './ui/DropdownMenu'
+import { DropdownMenu } from '../../../components/ui/DropdownMenu'
 import { EditStartupScriptModal } from './EditStartupScriptModal'
 import { EditProjectModal } from './EditProjectModal'
