+++ web/src/features/projects/components/ProjectManager.tsx
@@ -1,9 +1,9 @@
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { useState } from 'react'
-import { projectsApi, sessionsApi } from '../api/client'
-import { useAppStore } from '../store'
-import { Project } from '../types'
-import { DropdownMenu } from './ui/DropdownMenu'
+import { projectsApi, sessionsApi } from '../../../api/client'
+import { useAppStore } from '../../../store'
+import { Project } from '../../../types'
+import { DropdownMenu } from '../../../components/ui/DropdownMenu'
 import { CreateProjectModal } from './CreateProjectModal'
 import { CreateSSHProjectModal } from './CreateSSHProjectModal'
 import { EditProjectModal } from './EditProjectModal'
