--- web/src/features/projects/components/index.ts
@@ -1,4 +1,0 @@
-export { ProjectListContainer } from './ProjectListContainer'
-export { ProjectListView } from './ProjectListView'
-export { CreateProjectForm } from './CreateProjectForm'
-export { AddSSHProjectForm } from './AddSSHProjectForm'
