+++ web/src/features/sessions/components/SessionManager.tsx
@@ -1,12 +1,12 @@
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { useState, useEffect } from 'react'
-import { sessionsApi } from '../api/client'
-import { useAppStore } from '../store'
-import { Session } from '../types'
-import { useSessionTodos } from '../hooks/useSessionTodos'
-import { wsClient } from '../api/websocket'
-import { DropdownMenu } from './ui/DropdownMenu'
-import { useRunStartupScriptMutation } from '../features/sessions/api/sessionsApi'
+import { sessionsApi } from '../../../api/client'
+import { useAppStore } from '../../../store'
+import { Session } from '../../../types'
+import { useSessionTodos } from '../../todos/hooks/useSessionTodos'
+import { wsClient } from '../../../api/websocket'
+import { DropdownMenu } from '../../../components/ui/DropdownMenu'
+import { useRunStartupScriptMutation } from '../../sessions/api/sessionsApi'
 import { CreateSessionModal } from './CreateSessionModal'
 
 // Component to show in-progress task for a session
