+++ web/src/features/sessions/components/SessionView.tsx
@@ -1,13 +1,13 @@
 import { useState, useEffect } from 'react'
 import ReactMarkdown from 'react-markdown'
 import remarkGfm from 'remark-gfm'
-import { useAppStore } from '../store'
-import { AssistantView } from './AssistantView'
-import { FileDiffs } from './FileDiffs'
+import { useAppStore } from '../../../store'
+import { AssistantView } from '../../assistant/components/AssistantView'
+import { FileDiffs } from '../../git/components/FileDiffs'
 import { ManageSession } from './ManageSession'
-import { Terminal } from './Terminal'
-import { useDiffStats } from '../hooks/useDiffStats'
-import { useSessionActivity } from '../hooks/useSessionActivity'
+import { Terminal } from '../../terminal/components/Terminal'
+import { useDiffStats } from '../../git/hooks/useDiffStats'
+import { useSessionActivity } from '../../../hooks/useSessionActivity'
 
 type TabType = 'assistant' | 'diffs' | 'terminal' | 'manage'
