+++ web/src/features/settings/components/DarkModeToggle.tsx
@@ -1,4 +1,4 @@
-import { useAppStore } from '../store'
+import { useAppStore } from '../../../store'
 
 export function DarkModeToggle() {
   const { darkMode, setDarkMode } = useAppStore()
