+++ web/src/features/terminal/hooks/useTerminalManager.ts
