+++ web/src/features/todos/components/TaskDrawer.tsx
@@ -1,4 +1,4 @@
-import { RightDrawer } from './ui/RightDrawer'
+import { RightDrawer } from '../../../components/ui/RightDrawer'
 import { TodoList } from './TodoList'
 
 interface TaskDrawerProps {
